Druid 1.5 - Prevent early turn in to Shana Liskia in freportw.
Shana Liskia now requires the character to have the Mind Crystal and Protected Living Brambles seeds before accepting the Crystal Fragments.
This is a bandaid to prevent overflagging if the seeds turn in to Niera Farbreeze in surefall was missed.

It also allows characters that have already overflagged to recover and still turn in the seeds to get the Mind Crystal and Protected Living Brambles seed.