Update Camin.lua
Add a workaround for destroying Arantir's Ring by accident, or losing it in general.
The only known method of resetting the global is by completing the epic causing you in an infinite loop never being able to receive the ring again. This will delete the global after handing him 1000platinum allowing you to hail Dargon to turn into Arantir so you can hail for new ring.